backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "anthropic>=0.75.0",
    "asyncpg>=0.31.0",
    "celery>=5.6.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.125.0",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "bcrypt==4.3.0",
    "argon2-cffi>=23.1.0",
    "passlib>=1.7.4",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydub>=0.25.1",
    "python-dateutil>=2.9.0.post0",
    "python-jose[cryptography]>=3.5.0",
    "python-json-logger>=4.0.0",
    "python-magic>=0.4.27",
    "pytz>=2025.2",
    "redis>=7.1.0",
    "reportlab>=4.4.6",
    "sqlalchemy>=2.0.45",
    "stripe>=14.1.0",
    "python-multipart>=0.0.21",
    "google-generativeai>=0.8.6",
    "grok>=6.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
